Revert "Converted all of the note variables in the allocators to uint16_t"
This reverts commit 7c19c56e77b5b6012aec96a250d3d792054405a3.

+++ controller/voice_allocator.h
@@ -20,13 +20,12 @@
 #ifndef CONTROLLER_VOICE_ALLOCATOR_H_
 #define CONTROLLER_VOICE_ALLOCATOR_H_
 
-#include <string.h>
 #include "avrlib/base.h"
 
 namespace ambika {
 
 class VoiceAllocator {
- public:
+ public: 
   VoiceAllocator() { }
   void Init(uint8_t size, bool cyclic_mode, uint8_t* pool, uint8_t* lru) {
     size_ = size;
@@ -35,19 +34,19 @@ class VoiceAllocator {
     cyclic_allocator_ = cyclic_mode ? 0 : 0xff;
     Clear();
   }
-
-  uint8_t NoteOn(uint16_t note);
-  uint8_t NoteOff(uint16_t note);
-  uint8_t Find(uint16_t note) const;
-  uint8_t FindActive(uint16_t note) const;
-
+  
+  uint8_t NoteOn(uint8_t note);
+  uint8_t NoteOff(uint8_t note);
+  uint8_t Find(uint8_t note) const;
+  uint8_t FindActive(uint8_t note) const;
+  
   void Clear();
   void ClearNotes();
   uint8_t size() const { return size_; }
 
  private:
   void Touch(uint8_t voice);
-
+   
   // Holds the indices of the voices sorted by most recent usage.
   uint8_t* pool_;
   uint8_t* lru_;
